+++ src/main/java/com/lithiumcraft/createresourcegeodes/config/WandMode.java
@@ -1,0 +1,10 @@
+package com.lithiumcraft.createresourcegeodes.config;
+
+public enum WandMode {
+    MOVE,
+    BREAK;
+
+    public WandMode next() {
+        return values()[(this.ordinal() + 1) % values().length];
+    }
+}